requirements.txt
django>3.2,<3.3
djangorestframework
channels==3.0.4
channels_redis